Enhance database schema with indices and default values, and implement migration testing.
- **Database Schema Updates**:
    - Added performance-optimizing indices to `WorkoutSetEntity` (on `exercise`, `sessionId`, and `timestamp`) and `ExerciseEntity` (on `equipment` and `primaryMuscleGroup`).
    - Applied explicit `defaultValue` constraints to all columns in `ExerciseEntity` to ensure schema consistency.
    - Refined migration logic by moving the `session` table `notes` column addition from `MIGRATION_8_9` to `MIGRATION_7_8`.

- **Migration & Testing**:
    - Introduced `MigrationTest` to provide automated validation of schema transitions from version 7 through 9.
    - Added `androidx.room:room-testing` dependency to the test suite.

- **Build & Configuration**:
    - Configured KSP to export Room schemas to the project's `schemas` directory for version tracking.
    - Updated the Room schema JSON for version 9 to reflect the new indices and default value constraints.
    - Minor whitespace cleanup in `WorkoutViewModel`.

+++ app/src/main/java/com/example/gymdiary3/screens/SessionSummaryScreen.kt
@@ -128,7 +128,7 @@ fun SessionSummaryScreen(
                             val uiState = exerciseUiStates[entry.first] ?: ExerciseUiState(entry.first, 0.0, "Stable", false, "", 0.0, 0.0)
                             var historicBest by remember { mutableStateOf(0.0) }
                             LaunchedEffect(entry.first, s.session.id) {
-                                historicBest = viewModel.getHistoricBest1RM(entry.first, s.session.id.toLong())
+                                historicBest = viewModel.getHistoricBest1RM(entry.first, s.session.id)
                             }
                             ExerciseSummaryCard(isVisible, uiState, entry.second, userSettings.weightUnit, historicBest)
                         }
